refactor(backend): extract token boundary
