+++ backend/src/token.rs
@@ -1,0 +1,184 @@
+use std::sync::Arc;
+
+use axum::{extract::State, http::StatusCode, response::IntoResponse, Json};
+use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine as _};
+use chrono::{DateTime, Duration, Utc};
+use hmac::{Hmac, KeyInit, Mac};
+use jsonwebtoken::{encode, Algorithm, EncodingKey, Header};
+use serde::{Deserialize, Serialize};
+use sha2::Sha256;
+use tracing::error;
+
+use crate::{
+    error::ApiError, require_authenticated, users::GameRole, AppState, MAX_DISPLAY_NAME_LENGTH,
+};
+
+type HmacSha256 = Hmac<Sha256>;
+
+#[derive(Debug, Deserialize)]
+pub(crate) struct TokenRequest {
+    room: String,
+    name: String,
+}
+
+impl TokenRequest {
+    fn validate(&self) -> Result<(), ApiError> {
+        if self.room.trim().is_empty() {
+            return Err(ApiError::BadRequest("`room` must not be empty".to_string()));
+        }
+
+        let nickname = self.name.trim();
+        if nickname.is_empty() {
+            return Err(ApiError::BadRequest("`name` must not be empty".to_string()));
+        }
+        if nickname.chars().count() > MAX_DISPLAY_NAME_LENGTH {
+            return Err(ApiError::BadRequest(format!(
+                "`name` must be at most {MAX_DISPLAY_NAME_LENGTH} characters"
+            )));
+        }
+
+        Ok(())
+    }
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub(crate) struct TokenResponse {
+    pub(crate) token: String,
+    pub(crate) expires_at: DateTime<Utc>,
+    pub(crate) identity: String,
+    pub(crate) game_role: GameRole,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub(crate) struct LiveKitClaims {
+    pub(crate) iss: String,
+    pub(crate) sub: String,
+    pub(crate) name: String,
+    pub(crate) nbf: i64,
+    pub(crate) exp: i64,
+    pub(crate) video: LiveKitVideoGrant,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub(crate) struct LiveKitVideoGrant {
+    pub(crate) room: String,
+    #[serde(rename = "roomJoin")]
+    pub(crate) room_join: bool,
+    #[serde(rename = "canPublish")]
+    pub(crate) can_publish: bool,
+    #[serde(rename = "canSubscribe")]
+    pub(crate) can_subscribe: bool,
+}
+
+pub(crate) async fn mint_token(
+    State(state): State<Arc<AppState>>,
+    jar: axum_extra::extract::cookie::CookieJar,
+    Json(req): Json<TokenRequest>,
+) -> Result<impl IntoResponse, ApiError> {
+    req.validate()?;
+    let user = require_authenticated(&state, &jar).await?;
+
+    if let Some(allowed_rooms) = &state.config.allowed_rooms {
+        if !allowed_rooms.contains(&req.room) {
+            return Err(ApiError::RoomNotAllowed(req.room.clone()));
+        }
+    }
+
+    let google_subject = user.google_subject.clone().ok_or_else(|| {
+        ApiError::Forbidden("authenticated user is missing Google identity".to_string())
+    })?;
+    let opaque_identity = derive_room_identity(&state.config.cookie_secret, &google_subject)?;
+    let nickname = req.name.trim();
+
+    let updated_user = state
+        .user_store
+        .update_user_display_name(user.id, Some(nickname))
+        .await
+        .map_err(|err| {
+            error!("update display name error: {err}");
+            ApiError::Internal
+        })?;
+
+    let now = Utc::now();
+    let expiry = now + Duration::seconds(state.config.token_ttl_seconds as i64);
+
+    let claims = LiveKitClaims {
+        iss: state.config.livekit_api_key.clone(),
+        sub: opaque_identity,
+        name: updated_user
+            .display_name
+            .unwrap_or_else(|| nickname.to_string()),
+        nbf: now.timestamp(),
+        exp: expiry.timestamp(),
+        video: LiveKitVideoGrant {
+            room_join: true,
+            room: req.room,
+            can_publish: true,
+            can_subscribe: true,
+        },
+    };
+
+    let token = encode(
+        &Header::new(Algorithm::HS256),
+        &claims,
+        &EncodingKey::from_secret(state.config.livekit_api_secret.as_bytes()),
+    )
+    .map_err(|_| ApiError::Internal)?;
+
+    Ok((
+        StatusCode::OK,
+        Json(TokenResponse {
+            token,
+            expires_at: expiry,
+            identity: claims.sub,
+            game_role: updated_user.game_role,
+        }),
+    ))
+}
+
+pub(crate) fn derive_room_identity(secret: &str, google_subject: &str) -> Result<String, ApiError> {
+    let mut mac = HmacSha256::new_from_slice(secret.as_bytes()).map_err(|_| ApiError::Internal)?;
+    mac.update(b"livekit-identity:");
+    mac.update(google_subject.as_bytes());
+    Ok(format!(
+        "u_{}",
+        URL_SAFE_NO_PAD.encode(mac.finalize().into_bytes())
+    ))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn token_request_rejects_empty_room_or_name() {
+        assert!(matches!(
+            TokenRequest {
+                room: " ".to_string(),
+                name: "Alice".to_string(),
+            }
+            .validate(),
+            Err(ApiError::BadRequest(_))
+        ));
+        assert!(matches!(
+            TokenRequest {
+                room: "table".to_string(),
+                name: " ".to_string(),
+            }
+            .validate(),
+            Err(ApiError::BadRequest(_))
+        ));
+    }
+
+    #[test]
+    fn room_identity_is_stable_and_opaque() {
+        let first = derive_room_identity("secret", "google-subject").unwrap();
+        let second = derive_room_identity("secret", "google-subject").unwrap();
+        let different_subject = derive_room_identity("secret", "other-subject").unwrap();
+
+        assert_eq!(first, second);
+        assert_ne!(first, different_subject);
+        assert!(first.starts_with("u_"));
+        assert!(!first.contains("google-subject"));
+    }
+}
